fix: separation of methods of strategy on different files for better organization
