+++ meson.build
@@ -35,7 +35,9 @@ sources = files(
     'src/main.cpp',
 
     'src/pbts/control.cpp',
-    'src/pbts/strategy.cpp',
+    'src/pbts/strategy_r.cpp',
+    'src/pbts/strategy_d.cpp',
+    'src/pbts/utility.cpp',
     'src/pbts/simulator_connection.cpp',
     'src/pbts/common.cpp',
     'src/pbts/parse.cpp',
